refactor(cli): 🔨 move CheckDomainSettings to its own file
